Renaming files and creating Readme.md
