+++ headers/global_variables.h
